+++ src/main/java/org/xjcraft/hackerchecker/config/WarningConfig.java
@@ -3,11 +3,8 @@ package org.xjcraft.hackerchecker.config;
 import lombok.Data;
 import org.xjcraft.annotation.Instance;
 import org.xjcraft.annotation.RConfig;
-import org.xjcraft.hackerchecker.bean.MinerLog;
 
-import java.util.ArrayList;
 import java.util.LinkedHashMap;
-import java.util.List;
 import java.util.Map;
 
 @RConfig(configName = "warning.yml")
@@ -15,7 +12,7 @@ import java.util.Map;
 public class WarningConfig {
     @Instance
     public static WarningConfig config = new WarningConfig();
-    Map<String,Float> suspects = new LinkedHashMap<>();
+    Map<String, Double> suspects = new LinkedHashMap<>();
